Update README.md
 we made this project was to use a Raspberry Pi to design a home security
system. Hence, even when people are away from home, it helps them feel secure
about their home. This project is entirely built on the most recent Raspberry Pi 3b+
model, and Python is the coding language. Overall costs are modest, and operation is
simple. It will constantly engage with the user in an effort to improve energy
management and fully automate the home to guarantee privacy, security, and comfort.
On cell phones, the user can receive warnings via email at any time and from any
location. As a result, the system's architecture effectively blocks access for any
stranger trying to enter the house.